+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Main
+* [4400](https://github.com/grafana/loki/pull/4400) **trevorwhitney**: Config: automatically apply memberlist config too all rings when provided
 
 # 2.3.0 (2021/08/06)